fix save file and read file in same thread

+++ lib/savefile/save_csv.dart
@@ -3,109 +3,78 @@ import 'package:path_provider/path_provider.dart' as path_provider;
 import 'package:open_file/open_file.dart' as op;
 import 'package:excel/excel.dart';
 
-class SaveCsv {
-  String fileName;
-  List<String> header;
-  List<List<String>> body;
 
-  SaveCsv({this.fileName, this.header, this.body});
-
-  Future<String> get _localPath async {
-    final directory = File('/storage/emulated/0/Download/'); // await path_provider.getExternalStorageDirectory();
-    return directory.path;
-  }
-
-  Future<File> get _localFile async {
-    final path = await _localPath;
-    return File('$path/${fileName}.xlsx');
-  }
-
-  Future<void> writeCSV() async {
-    write().then((file) async {
-      String path = file.path;
-      await op.OpenFile.open(path);
-    });
-  }
-
-
-  Future<File> write() async {
-    final file = await _localFile;
-
-    if (file.existsSync()){
-      file.deleteSync(recursive: true);
-    }
-
-    String doc = "";
-    String h = header.join(', ');
-    doc += "${h}\n";
-
-    for (var b in body){
-        String sb = b.join(', ');
-        doc += "${sb}\n";
-    }
-    await file.writeAsString("${doc}",mode: FileMode.append, flush: true);
-
-    return file;
-  }
-}
-
-
+// ini adalah class yang memiliki
+// fungsi menyimpan ke dalam file
+// excel
 class SaveExcel {
+
+  // nama file
   String fileName;
+
+  // nama kolom
   List<String> header;
+
+  // isi dokumen
   List<List<String>> body;
 
   SaveExcel({this.fileName, this.header, this.body});
 
+  // mengambil alamat lokasi penyimpanan
   Future<String> get _localPath async {
     final directory = File('/storage/emulated/0/Download/'); // await path_provider.getExternalStorageDirectory();
     return directory.path;
   }
 
+  // membuat file
   Future<File> get _localFile async {
     final path = await _localPath;
     return File('$path/${fileName}.xlsx');
   }
 
+  // tuliskan ke dalam format excel
   Future<void> writeExcel() async {
-    write().then((file) async {
+    final file = await _localFile;
+
+    // jika ada, hapus dulu
+    if (file.existsSync()){ file.deleteSync(recursive: true); }
+
+    write().then((onValue) {
+      file
+        ..createSync(recursive: true)
+        ..writeAsBytesSync(onValue);
+
       String path = file.path;
-      await op.OpenFile.open(path);
+
+      // membuka file excel yang telah dibuat
+      op.OpenFile.open(path);
     });
   }
 
+  // fungsi untuk mengisi file excel
+  Future<List<int>> write() async {
 
-  Future<File> write() async {
-    final file = await _localFile;
-
-    if (file.existsSync()){
-      file.deleteSync(recursive: true);
-    }
-
+    // buat instance excel
     var excel = Excel.createExcel();
     var sheet = excel['Sheet1'];
 
+    // generate cell
     List<List<String>> list = List.generate(body.length + 1, (index) => List.generate(header.length , (index1) => '$index $index1'));
-    list.forEach((row) {
-      sheet.appendRow(row);
-    });
+    list.forEach((row) { sheet.appendRow(row); });
 
+    // isi baris pertama dengan nama kolom
     for (int i = 0; i < header.length; i++) {
       sheet.row(0)[i].value = header[i];
     }
 
+    // dan sisanya dengan data dari body
     for (int i = 0;i < body.length; i++) {
       for (int o = 0; o < body[i].length; o++) {
         sheet.row(i + 1)[o].value = body[i][o];
       }
     }
 
-    excel.encode().then((onValue) {
-      file
-        ..createSync(recursive: true)
-        ..writeAsBytesSync(onValue);
-    });
-
-    return file;
+    // encode ke format excel
+    return excel.encode();
   }
 }
